Create nightly test CI and add invite expire test to it
