+++ backend/test/test_invites.py
@@ -41,48 +41,3 @@ def test_pending_invites_crawler(crawler_auth_headers, default_org_id):
     # Verify that only superusers can see pending invites
     r = requests.get(f"{API_PREFIX}/users/invites", headers=crawler_auth_headers)
     assert r.status_code == 403
-
-
-def test_invites_expire(admin_auth_headers, default_org_id):
-    """Note this test is dependent on chart/test/test.yaml settings.
-
-    Namely, it expects `invite_expire_seconds: 10` to be set in chart.
-    """
-    # Send invite
-    INVITE_EMAIL = "invite-expires@example.com"
-    r = requests.post(
-        f"{API_PREFIX}/orgs/{default_org_id}/invite",
-        headers=admin_auth_headers,
-        json={"email": INVITE_EMAIL, "role": 10},
-    )
-    assert r.status_code == 200
-    data = r.json()
-    assert data["invited"] == "new_user"
-
-    # Verify invite exists
-    r = requests.get(
-        f"{API_PREFIX}/orgs/{default_org_id}/invites",
-        headers=admin_auth_headers,
-    )
-    assert r.status_code == 200
-    data = r.json()
-    invites_matching_email = [
-        invite for invite in data["pending_invites"] if invite["email"] == INVITE_EMAIL
-    ]
-    assert len(invites_matching_email) == 1
-
-    # Wait 90 seconds to give Mongo time to delete the invite
-    # See: https://www.mongodb.com/docs/manual/core/index-ttl/#timing-of-the-delete-operation
-    time.sleep(90)
-
-    # Check invites again and verify invite has been removed
-    r = requests.get(
-        f"{API_PREFIX}/orgs/{default_org_id}/invites",
-        headers=admin_auth_headers,
-    )
-    assert r.status_code == 200
-    data = r.json()
-    invites_matching_email = [
-        invite for invite in data["pending_invites"] if invite["email"] == INVITE_EMAIL
-    ]
-    assert len(invites_matching_email) == 0
