+++ backend/test_nightly/conftest.py
@@ -1,0 +1,43 @@
+import pytest
+import requests
+import time
+
+
+HOST_PREFIX = "http://127.0.0.1:30870"
+API_PREFIX = HOST_PREFIX + "/api"
+
+ADMIN_USERNAME = "admin@example.com"
+ADMIN_PW = "PASSW0RD!"
+
+
+@pytest.fixture(scope="session")
+def admin_auth_headers():
+    while True:
+        r = requests.post(
+            f"{API_PREFIX}/auth/jwt/login",
+            data={
+                "username": ADMIN_USERNAME,
+                "password": ADMIN_PW,
+                "grant_type": "password",
+            },
+        )
+        data = r.json()
+        try:
+            return {"Authorization": f"Bearer {data['access_token']}"}
+        except:
+            print("Waiting for admin_auth_headers")
+            time.sleep(5)
+
+
+@pytest.fixture(scope="session")
+def default_org_id(admin_auth_headers):
+    while True:
+        r = requests.get(f"{API_PREFIX}/orgs", headers=admin_auth_headers)
+        data = r.json()
+        try:
+            for org in data["orgs"]:
+                if org["default"] is True:
+                    return org["id"]
+        except:
+            print("Waiting for default org id")
+            time.sleep(5)
